move post endpoint to /[id]
