+++ src/pages/api/title/[id].ts
@@ -1,5 +1,6 @@
 import {NextApiRequest, NextApiResponse} from 'next';
 import prisma from '@/lib/prisma';
+import {title} from '@prisma/client';
 
 
 // ANY api/title/[id]
@@ -12,6 +13,8 @@ export default async function handle(
   switch (req.method) {
   case 'GET':
     return handleGET(titleId as string, res);
+  case 'POST':
+    return handlePOST(req.body as title, res);
   default:
     throw new Error('Method not supported');
   }
@@ -28,4 +31,17 @@ async function handleGET(titleId: string, res: NextApiResponse) {
   });
 
   return res.status(200).json(title);
+}
+
+// POST api/title/[id]
+async function handlePOST(localTitle: title, res: NextApiResponse) {
+  await prisma.title.upsert({
+    where: { id: localTitle.id },
+    update: { ...localTitle },
+    create: { ...localTitle },
+  }).catch(e => {
+    throw new Error(`Failed to create or update title: ${e}`);
+  });
+
+  return res.status(200).json(localTitle);
 }
